Unify workspace surface definitions
